fix: include encrypted value in create --secret payload
When using `reforge create --secret`, the encrypted value was missing
from the API payload. The `mapConfigValueToDto` function wasn't
handling the `value` property returned by `makeConfidentialValue`.

Changes:
- Extract `mapConfigValueToDto` to shared utility (config-value-dto.ts)
- Add handling for `configValue.value` (used by encrypted values)
- Update both create.ts and set-default.ts to use shared utility
- Add payload validation in create test mock for encrypted values
- Add `set-default --secret` to interactive menu for consistency
- Bump version to 0.0.14
